requirements.txt
flask
cohere
gunicorn
scikit-learn
python-dotenv